+ Change the way create Puzzle (now it is sure that have the way) + Fix gameLogic so that food cannot eat wall (not sure enough) + KeyPress now return value so that can draw backgound when need + Point now is float + Change something in constant

+++ Bonus.h
@@ -6,7 +6,7 @@ class Bonus : public Obstacle
 {
 public:
 	void Random();
-	void CollidePoccess(int& point, Ball& ball);
+	void CollidePoccess(float& point, Ball& ball);
 
 public:
 	Bonus();
